+++ WonderMake/Utilities/Vector.h
@@ -294,10 +294,10 @@ struct SVector
 	template<u64 TArraySize, typename TVector> requires std::is_same_v<SVector, std::decay_t<TVector>>
 	[[nodiscard]] static constexpr TVector& Closest(SVector aPoint, std::array<TVector*, TArraySize> aArray) noexcept
 	{
-		u64 index = 0;
+		size_t index = 0;
 		f32 distance = std::numeric_limits<f32>::max();
 
-		for (u64 i = 0; i < aArray.size(); i++)
+		for (size_t i = 0; i < aArray.size(); i++)
 		{
 			auto testVector = *aArray[i];
 			const f32 newDistance = aPoint.DistanceTo(testVector);
